+++ network-module/src/main/java/it/jnrpe/network/encoders/V4Encoder.java
@@ -20,19 +20,18 @@ class V4Encoder implements IResponseEncoder {
         DataOutputStream dout = new DataOutputStream(bout);
 
         try {
-            dout.writeShort(response.getVersion()); // Version 4
+            dout.writeShort(response.getVersion()); // Version 3
             dout.writeShort(response.getPacketType()); // Type: Response
             dout.writeInt((int) response.getCrc32()); // Type: Response
             dout.writeShort(response.getResultCode());
             dout.writeShort(response.getAlignment());
-            dout.writeShort(response.getBuffer().length);
+            dout.writeInt(response.getBuffer().length);
             dout.write(response.getBuffer());
             dout.write(response.getPadding());
             dout.flush();
         } catch (Exception e) {
 
         }
-
         return bout.toByteArray();
     }
 }
